设计模式

+++ auth-common/src/main/java/com/platform/authcommon/common/ErrorCode.java
@@ -5,7 +5,7 @@ package com.platform.authcommon.common;
  */
 public interface ErrorCode {
 
-    long getCode();
+    Integer getCode();
 
     String getMessage();
 }
