+++ auth-common/src/main/java/com/platform/authcommon/common/ResultCode.java
@@ -11,15 +11,15 @@ public enum ResultCode implements ErrorCode {
     UNAUTHORIZED(401, "暂未登录或token已经过期"),
     FORBIDDEN(403, "没有相关权限");
 
-    private long code;
+    private Integer code;
     private String message;
 
-    ResultCode(long code, String message) {
+    ResultCode(Integer code, String message) {
         this.code = code;
         this.message = message;
     }
 
-    public long getCode() {
+    public Integer getCode() {
         return code;
     }
